refactor reset-password page
